style: Rename PuzzleSet files
