+++ frontend/src/components/puzzle_sets/PS1P2.tsx
